Volgende dingen toegevoegd: - Annotaties toegevoegd - Verschillende activities verwijderd die niet meer nodig waren
