+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
 
         </activity>
 
-        <activity android:name=".PlayActivity" >
+        <activity android:name=".MainActivity" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
@@ -21,10 +21,6 @@
         </activity>
         <activity android:name=".ResultActivity" />
         <activity android:name=".QuizActivity" />
-        <activity android:name=".ExerciseActivity" />
-        <activity android:name=".ProgressActivity" />
-        <activity android:name=".SettingsActivity" />
-        <activity android:name=".MainActivity"/>
         <service android:name=".MyFirebaseMessagingService"
                  android:exported="false">
             <intent-filter>
@@ -32,5 +28,4 @@
             </intent-filter>
         </service>
     </application>
-
 </manifest>
