--- app/src/main/java/com/example/iiatimd_project_1920/ExerciseActivity.java
@@ -1,14 +1,0 @@
-package com.example.iiatimd_project_1920;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class ExerciseActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_exercise);
-    }
-}
