+++ app/src/main/java/com/example/iiatimd_project_1920/QuestionViewModel.java
@@ -10,16 +10,13 @@ import java.util.List;
 public class QuestionViewModel extends AndroidViewModel {
 
     private QuestionsRepository mRepository;
-
     private LiveData<List<Questions>> mAllQuestions;
 
     public QuestionViewModel(Application application){
         super(application);
         mRepository = new QuestionsRepository(application);
         mAllQuestions = mRepository.getmAllQuestions();
-
     }
-
     LiveData<List<Questions>> getmAllQuestions()
     {
         return mAllQuestions;
