+++ app/src/main/java/com/example/iiatimd_project_1920/SplashActivity.java
@@ -71,7 +71,7 @@ public class SplashActivity extends AppCompatActivity {
     }
 
     private void GoToPlayActivity() {
-        startActivityForResult(new Intent(SplashActivity.this, PlayActivity.class), EXIT_CODE);
+        startActivityForResult(new Intent(SplashActivity.this, MainActivity.class), EXIT_CODE);
     }
 
     @Override
