+++ app/src/main/java/com/example/iiatimd_project_1920/TimerDialog.java
@@ -5,7 +5,6 @@ import android.content.Context;
 import android.content.Intent;
 import android.view.View;
 import android.widget.Button;
-import android.widget.TextView;
 
 public class TimerDialog {
 
@@ -31,7 +30,7 @@ public class TimerDialog {
             @Override
             public void onClick(View v) {
                 finalScoreDialog.dismiss();
-                Intent intent = new Intent(mContext, PlayActivity.class);
+                Intent intent = new Intent(mContext, MainActivity.class);
                 mContext.startActivity(intent);
             }
         });
